Added and organized base 'signin' logic
